apps/controller/requirements.txt
jsonschema==4.17.3
kubernetes==25.3.0
kafka-python==2.0.2
APScheduler==3.10.1
requests==2.31.0